+++ tests/HandlerTest.php
@@ -12,7 +12,7 @@ use PHPUnit\Framework\TestCase;
 
 class HandlerTest extends TestCase
 {
-    private static function getContainer(): WhoopsHandlerContainer
+    private static function getHttpContainer(): WhoopsHandlerContainer
     {
         return new class() extends WhoopsHandlerContainer {
             protected static function isCli(): bool
@@ -22,12 +22,22 @@ class HandlerTest extends TestCase
         };
     }
 
+    private static function getCliContainer(): WhoopsHandlerContainer
+    {
+        return new class() extends WhoopsHandlerContainer {
+            protected static function isCli(): bool
+            {
+                return true;
+            }
+        };
+    }
+
     public function testJson(): void
     {
         $request = Factory::createServerRequest('GET', '/')->withHeader('Accept', 'application/json');
 
         $response = Dispatcher::run([
-            (new Whoops())->handlerContainer(self::getContainer()),
+            (new Whoops())->handlerContainer(self::getHttpContainer()),
             function () {
                 throw new Exception('Error Processing Request');
             },
@@ -37,27 +47,51 @@ class HandlerTest extends TestCase
         $this->assertEquals('application/json', $response->getHeaderLine('Content-Type'));
     }
 
-    public function testXml(): void
+    public function dataProviderForXml(): array
     {
-        $request = Factory::createServerRequest('GET', '/')->withHeader('Accept', 'text/xml');
+        return [
+            ['text/xml'],
+            ['application/xml'],
+        ];
+    }
+
+    /**
+     * @dataProvider dataProviderForXml
+     */
+    public function testXml(string $accept): void
+    {
+        $request = Factory::createServerRequest('GET', '/')->withHeader('Accept', $accept);
 
         $response = Dispatcher::run([
-            (new Whoops())->handlerContainer(self::getContainer()),
+            (new Whoops())->handlerContainer(self::getHttpContainer()),
             function () {
                 throw new Exception('Error Processing Request');
             },
         ], $request);
 
         $this->assertEquals(500, $response->getStatusCode());
-        $this->assertEquals('text/xml', $response->getHeaderLine('Content-Type'));
+        $this->assertEquals('application/xml', $response->getHeaderLine('Content-Type'));
     }
 
-    public function testPlain(): void
+    public function dataProviderForPlain(): array
     {
-        $request = Factory::createServerRequest('GET', '/')->withHeader('Accept', 'text/plain');
+        return [
+            ['text/plain'],
+            ['text/css'],
+            ['text/javascript; charset=utf-8'],
+            ['application/javascript'],
+        ];
+    }
+
+    /**
+     * @dataProvider dataProviderForPlain
+     */
+    public function testPlain(string $accept): void
+    {
+        $request = Factory::createServerRequest('GET', '/')->withHeader('Accept', $accept);
 
         $response = Dispatcher::run([
-            (new Whoops())->handlerContainer(self::getContainer()),
+            (new Whoops())->handlerContainer(self::getHttpContainer()),
             function () {
                 throw new Exception('Error Processing Request');
             },
@@ -72,7 +106,7 @@ class HandlerTest extends TestCase
         $request = Factory::createServerRequest('GET', '/')->withHeader('Accept', 'text/html');
 
         $response = Dispatcher::run([
-            (new Whoops())->handlerContainer(self::getContainer()),
+            (new Whoops())->handlerContainer(self::getHttpContainer()),
             function () {
                 throw new Exception('Error Processing Request');
             },
@@ -87,7 +121,7 @@ class HandlerTest extends TestCase
         $request = Factory::createServerRequest('GET', '/')->withHeader('Accept', 'foo/bar');
 
         $response = Dispatcher::run([
-            (new Whoops())->handlerContainer(self::getContainer()),
+            (new Whoops())->handlerContainer(self::getHttpContainer()),
             function () {
                 throw new Exception('Error Processing Request');
             },
@@ -100,7 +134,7 @@ class HandlerTest extends TestCase
     public function testEmptyAccept(): void
     {
         $response = Dispatcher::run([
-            (new Whoops())->handlerContainer(self::getContainer()),
+            (new Whoops())->handlerContainer(self::getHttpContainer()),
             function () {
                 throw new Exception('Error Processing Request');
             },
@@ -108,5 +142,20 @@ class HandlerTest extends TestCase
 
         $this->assertEquals(500, $response->getStatusCode());
         $this->assertEquals('text/html', $response->getHeaderLine('Content-Type'));
+    }
+
+    public function testCli(): void
+    {
+        $request = Factory::createServerRequest('GET', '/')->withHeader('Accept', 'foo/bar');
+
+        $response = Dispatcher::run([
+            (new Whoops())->handlerContainer(self::getCliContainer()),
+            function () {
+                throw new Exception('Error Processing Request');
+            },
+        ], $request);
+
+        $this->assertEquals(500, $response->getStatusCode());
+        $this->assertEquals('text/plain', $response->getHeaderLine('Content-Type'));
     }
 }